Limpieza y refactorizacion de código

+++ app/src/main/java/com/alaturing/incidentify/main/incident/data/remote/IncidentRemoteDatasource.kt
@@ -12,7 +12,7 @@ interface IncidentRemoteDatasource {
      * @return Resultado de lista de incidentes
      */
     suspend fun createOne(
-        localId:Int,
+        localId:Long,
         description:String,
                           evidence: Uri?,
                           latitude:Double?,
