+++ app/src/main/java/com/alaturing/incidentify/main/incident/model/Incident.kt
@@ -6,8 +6,8 @@ import android.net.Uri
  * Entidad modelo que representa un incidente
  */
 data class Incident(
-    val localId:Int,
-    val remoteId:Int?,
+    val localId:Long,
+    val remoteId:Long?,
     val description:String,
     val solved:Boolean,
     val solved_at:Long? = null,
