+++ app/src/main/java/com/alaturing/incidentify/main/incident/ui/edit/IncidentEditUiState.kt
@@ -2,7 +2,7 @@ package com.alaturing.incidentify.main.incident.ui.edit
 
 sealed class IncidentEditUiState {
     data object New: IncidentEditUiState()
-    data class Created(val id:Int): IncidentEditUiState()
+    data class Created(val id:Long): IncidentEditUiState()
     data class Error(val message:String): IncidentEditUiState()
 
 }
